+++ src/clib/cool1d_multi_g.cpp
@@ -27,7 +27,7 @@
 #include "dust_props.hpp"
 #include "inject_model/grain_metal_inject_pathways.hpp"
 #include "internal_types.hpp"
-#include "interp_table_utils.hpp"
+#include "interp_grid.hpp"
 #include "opaque_storage.hpp"
 #include "tabulated/cool1d_cloudy.hpp"
 #include "tabulated/cool1d_cloudy_old_tables.hpp"
